# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SecureShare Authorization Reference API",
    "summary": "Multi-tenant document sharing with JWT, OpenFGA, Macaroons, PostgreSQL, and audit logging.",
    "description": "\nSecureShare is a production-style reference API for secure multi-tenant document sharing.\n\nCore security properties:\n\n* JWTs carry identity only, never document permissions.\n* FastAPI routers and dependencies act as policy enforcement points.\n* OpenFGA is the relationship-based policy decision point for document access.\n* Macaroon-style delegated links provide temporary, attenuated read access.\n* Every allow and deny authorization decision is written to the audit log.\n\nDemo users all use password `password123`:\n\n* `alice@example.com` owns the Acme document.\n* `bob@example.com` can view the Acme document.\n* `eve@example.net` owns the Globex document in another tenant.\n",
    "contact": {
      "name": "SecureShare maintainers"
    },
    "license": {
      "name": "Reference implementation"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/documents/{document_id}/shares": {
      "post": {
        "tags": [
          "shares"
        ],
        "summary": "Grant document access",
        "description": "Grants `owner`, `editor`, `commenter`, or `viewer` on a document to a user or group. The caller must pass the shared authorization service check for `can_share`.",
        "operationId": "share_document_documents__document_id__shares_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShareRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Relationship granted in OpenFGA."
          },
          "401": {
            "description": "Missing or invalid JWT."
          },
          "403": {
            "description": "Caller cannot share this document."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "shares"
        ],
        "summary": "Revoke document access",
        "description": "Revokes a document relationship from a user or group in OpenFGA. Revocation takes effect on the next request without requiring JWT refresh.",
        "operationId": "revoke_document_share_documents__document_id__shares_delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShareRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Relationship revoked in OpenFGA."
          },
          "401": {
            "description": "Missing or invalid JWT."
          },
          "403": {
            "description": "Caller cannot share this document."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/documents/{document_id}/shares/delegated-link": {
      "post": {
        "tags": [
          "shares"
        ],
        "summary": "Create a delegated read link",
        "description": "Issues a Macaroon-style delegated read token with caveats for action, document, tenant, issuer, expiry, and caller IP address. The issuer must currently have `can_read`.",
        "operationId": "create_delegated_link_documents__document_id__shares_delegated_link_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DelegatedLinkRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Delegated token issued.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DelegatedLinkResponse"
                }
              }
            }
          },
          "400": {
            "description": "Caller IP could not be bound into the delegated token."
          },
          "401": {
            "description": "Missing or invalid JWT."
          },
          "403": {
            "description": "Issuer cannot read this document."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DelegatedLinkRequest": {
        "properties": {
          "expires_in_seconds": {
            "type": "integer",
            "maximum": 604800.0,
            "exclusiveMinimum": 0.0,
            "title": "Expires In Seconds"
          },
          "ip_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ip Address"
          }
        },
        "type": "object",
        "required": [
          "expires_in_seconds"
        ],
        "title": "DelegatedLinkRequest",
        "examples": [
          {
            "expires_in_seconds": 300,
            "ip_address": "203.0.113.10"
          },
          {
            "expires_in_seconds": 900
          }
        ]
      },
      "DelegatedLinkResponse": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          },
          "caveats": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Caveats"
          }
        },
        "type": "object",
        "required": [
          "token",
          "caveats"
        ],
        "title": "DelegatedLinkResponse",
        "examples": [
          {
            "caveats": [
              "action = can_read",
              "document_id = aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa",
              "expires_before = 1893456000"
            ],
            "token": "MDAxZGxvY2F0aW9u..."
          }
        ]
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ShareRequest": {
        "properties": {
          "subject_type": {
            "type": "string",
            "pattern": "^(user|group)$",
            "title": "Subject Type"
          },
          "subject_id": {
            "type": "string",
            "title": "Subject Id"
          },
          "role": {
            "type": "string",
            "pattern": "^(editor|commenter|viewer)$",
            "title": "Role"
          }
        },
        "type": "object",
        "required": [
          "subject_type",
          "subject_id",
          "role"
        ],
        "title": "ShareRequest",
        "examples": [
          {
            "role": "viewer",
            "subject_id": "00000000-0000-0000-0000-000000000002",
            "subject_type": "user"
          },
          {
            "role": "commenter",
            "subject_id": "33333333-3333-3333-3333-333333333333",
            "subject_type": "group"
          }
        ]
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}